Make TestJobRunner() platform independant
